\name{test_gatkreport.table}
\docType{data}
\alias{test_gatkreport_v1.table}
\title{Test table for version 1.x GATKReport}
\description{
   This is a new-style GATKReport.
}
\usage{test_gatkreport.table}
\format{Text document containing multiple tables}
\source{GATK test data, Broad Institute}
\references{
   https://gatk.broadinstitute.org/hc/en-us/articles/360035532172-GATKReport-and-gsalib
}
