\name{test_genconcord.table}
\docType{data}
\alias{test_genconcord.table}
\title{Test table for Genotype Concordance Table Reshape}
\description{
   This is a GATKReport output by GenotypeConcordance.
}
\usage{test_genconcord.table}
\format{Text document containing multiple tables}
\source{GATK test data, Broad Institute}
\references{
   https://gatk.broadinstitute.org/hc/en-us/articles/360035532172-GATKReport-and-gsalib
}
